# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/deployments:
    get:
      description: Gets all deployments.
      operationId: getDeployments
      parameters:
        - name: page
          required: false
          in: query
          description: The page count.
          schema:
            type: number
        - name: pageSize
          required: false
          in: query
          description: The page size.
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentsDto"
      security:
        - x-api-key: []
      summary: ""
      tags:
        - deployments
  /api/deployments/{deploymentId}:
    get:
      description: Gets a deployment by ID.
      operationId: getDeployment
      parameters:
        - name: deploymentId
          required: true
          in: path
          description: The ID of the deployment.
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentDto"
      security:
        - x-api-key: []
      summary: ""
      tags:
        - deployments
    put:
      description: Updates a deployment.
      operationId: putDeployment
      parameters:
        - name: deploymentId
          required: true
          in: path
          description: The ID of the deployment.
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeploymentDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentDto"
      security:
        - x-api-key: []
      summary: ""
      tags:
        - deployments
    delete:
      description: Delete a deployment.
      operationId: deleteDeployment
      parameters:
        - name: deploymentId
          required: true
          in: path
          description: The ID of the deployment.
          schema:
            type: number
      responses:
        "204":
          description: ""
      security:
        - x-api-key: []
      summary: ""
      tags:
        - deployments
  /api/deployments/{deploymentId}/retry:
    post:
      description: Retries a failed deployment.
      operationId: retryDeployment
      parameters:
        - name: deploymentId
          required: true
          in: path
          description: The ID of the deployment.
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentDto"
      security:
        - x-api-key: []
      summary: ""
      tags:
        - deployments
  /api/deployments/{deploymentId}/status:
    get:
      description: Gets deployments status.
      operationId: getDeploymentStatus
      parameters:
        - name: deploymentId
          required: true
          in: path
          description: The ID of the deployment.
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentStatusDto"
      security:
        - x-api-key: []
      summary: ""
      tags:
        - deployments
  /api/deployments/{deploymentId}/steps:
    get:
      description: Gets the steps of the most recent deployment update.
      operationId: getDeploymentSteps
      parameters:
        - name: deploymentId
          required: true
          in: path
          description: The ID of the deployment.
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentStepsDto"
      security:
        - x-api-key: []
      summary: ""
      tags:
        - deployments
  /api/deployments/{deploymentId}/logs:
    get:
      description: Gets deployments logs.
      operationId: getDeploymentLogs
      parameters:
        - name: deploymentId
          required: true
          in: path
          description: The ID of the deployment.
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentLogsDto"
      security:
        - x-api-key: []
      summary: ""
      tags:
        - deployments
  /api/deployments/{deploymentId}/checks:
    get:
      description: Gets deployments status.
      operationId: getDeploymentChecks
      parameters:
        - name: deploymentId
          required: true
          in: path
          description: The ID of the deployment.
          schema:
            type: number
        - name: fromDate
          required: true
          in: query
          schema:
            type: string
        - name: toDate
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentCheckSummariesDto"
      security:
        - x-api-key: []
      summary: ""
      tags:
        - deployments
  /api/deployments/{deploymentId}/usage:
    get:
      description: Gets usage summaries.
      operationId: getDeploymentUsage
      parameters:
        - name: deploymentId
          required: true
          in: path
          description: The ID of the deployment.
          schema:
            type: number
        - name: fromDate
          required: true
          in: query
          schema:
            type: string
        - name: toDate
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentUsageSummariesDto"
      security:
        - x-api-key: []
      summary: ""
      tags:
        - deployments
...
  /api/deployments/{deploymentId}/confirm:
    get:
      operationId: DeploymentsController_confirmDeployment
      parameters:
        - name: teamId
          required: true
          in: path
          schema:
            type: number
        - name: deploymentId
          required: true
          in: path
          schema:
            type: number
        - name: token
          required: true
          in: query
          schema:
            type: string
        - name: redirectUrl
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - x-api-key: []
      tags:
        - deployments
  /api/deployments/{deploymentId}/cancel:
    get:
      operationId: DeploymentsController_cancelDeployment
      parameters:
        - name: deploymentId
          required: true
          in: path
          schema:
            type: number
        - name: token
          required: true
          in: query
          schema:
            type: string
        - name: redirectUrl
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - x-api-key: []
      tags:
        - deployments
...
  /api/services/{serviceId}/deployments:
    get:
      description: Gets all deployments.
      operationId: getServiceDeployments
      parameters:
        - name: serviceId
          required: true
          in: path
          description: The ID of the service.
          schema:
            type: number
        - name: page
          required: false
          in: query
          description: The page count.
          schema:
            type: number
        - name: pageSize
          required: false
          in: query
          description: The page size.
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentsDto"
      security:
        - x-api-key: []
      summary: ""
      tags:
        - services
...
  /api/teams/{teamId}/deployments:
    get:
      description: Gets all deployments.
      operationId: getTeamDeployments
      parameters:
        - name: teamId
          required: true
          in: path
          description: The ID of the team.
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentsDto"
      security:
        - x-api-key: []
      summary: ""
      tags:
        - deployments
    post:
      description: Creates a deployment.
      operationId: postTeamDeployment
      parameters:
        - name: teamId
          required: true
          in: path
          description: The ID of the team.
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeploymentDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentCreatedDto"
      security:
        - x-api-key: []
      summary: ""
      tags:
        - deployments
...
info:
  title: Pangostack
  description: SaaS Deployment Tool
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    x-api-key:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    LabeledValueDto:
      type: object
      properties:
        value:
          type: string
          description: The value, for example a connection URL or an image version.
        isPublic:
          type: boolean
          description: Whether this value is publicly visible to the customer.
        label:
          type: string
          description: Human-readable label for the value.
      required:
        - value
        - isPublic
        - label
    AvailableUpdateDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the version.
      required:
        - name
    DeploymentResourceDto:
      type: object
      properties:
        id:
          type: string
          description: The ID of the resource.
        name:
          type: string
          description: The name of the resource.
      required:
        - id
        - name
    DeploymentDto:
      type: object
      properties:
        id:
          type: number
          description: The ID of the deployment.
        name:
          type: string
          description: The name of the deployment.
          nullable: true
        serviceId:
          type: number
          description: The ID of the service.
        serviceName:
          type: string
          description: The name of the service.
        serviceVersion:
          type: string
          description: The version of the service.
        isVersionDefault:
          type: boolean
          description: Indicates if the version is the default one.
        isVersionLast:
          type: boolean
          description: " Indicates if the version is the last one."
        createdAt:
          type: string
          description: The timestamp when the deployment has been created.
          format: date
        connections:
          type: object
          description: The connection infos organized by connection type and name.
          additionalProperties:
            type: object
            additionalProperties:
              $ref: "#/components/schemas/LabeledValueDto"
        afterInstallationInstructions:
          type: object
          description: Instructions to follow after installation.
          nullable: true
        status:
          type: string
          description: The current status of the last deployment update.
          enum:
            - Pending
            - Running
            - Completed
            - Failed
        healthStatus:
          type: string
          description: The current health status of the deployment.
          nullable: true
          enum:
            - Succeeded
            - Failed
        parameters:
          type: object
          description: The current deployment parameters.
          additionalProperties: true
        resources:
          description: The current resources.
          type: array
          items:
            $ref: "#/components/schemas/DeploymentResourceDto"
        availableUpdates:
          description: The available updates.
          type: array
          items:
            $ref: "#/components/schemas/AvailableUpdateDto"
      required:
        - id
        - name
        - serviceId
        - serviceName
        - serviceVersion
        - isVersionDefault
        - isVersionLast
        - createdAt
        - connections
        - afterInstallationInstructions
        - status
        - healthStatus
        - parameters
        - resources
        - availableUpdates
    DeploymentsDto:
      type: object
      properties:
        items:
          description: The deployments.
          type: array
          items:
            $ref: "#/components/schemas/DeploymentDto"
        total:
          type: number
          description: The total number of deployments.
      required:
        - items
        - total
    UpdateDeploymentDto:
      type: object
      properties:
        name:
          type: string
          description: The optional name to describe the deployment.
          nullable: true
          maxLength: 100
        versionId:
          type: number
          description: The ID of the version when an update is made.
          nullable: true
        parameters:
          type: object
          description: The parameters.
          additionalProperties: true
      required:
        - name
        - versionId
        - parameters
    ResourceNodeStatusDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the node.
        isReady:
          type: boolean
          description: Indicates if the node can be used.
        message:
          type: string
          description: The message to describe the status.
          nullable: true
      required:
        - name
        - isReady
        - message
    ResourceWorkloadStatusDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the workload.
        nodes:
          description: All nodes within the workload.
          type: array
          items:
            $ref: "#/components/schemas/ResourceNodeStatusDto"
      required:
        - name
        - nodes
    ResourceStatusDto:
      type: object
      properties:
        resourceId:
          type: string
          description: The resource I.
        resourceType:
          type: string
          description: The type of the resource.
        resourceName:
          type: string
          description: The name of the resource.
        workloads:
          description: The workflows that have been create.
          type: array
          items:
            $ref: "#/components/schemas/ResourceWorkloadStatusDto"
        properties:
          type: object
          description: Runtime properties collected live from the resource, for example
            the deployed image version.
          additionalProperties:
            $ref: "#/components/schemas/LabeledValueDto"
      required:
        - resourceId
        - resourceType
        - resourceName
        - workloads
        - properties
    DeploymentStatusDto:
      type: object
      properties:
        resources:
          description: The resource.
          type: array
          items:
            $ref: "#/components/schemas/ResourceStatusDto"
      required:
        - resources
    DeploymentStepLogDto:
      type: object
      properties:
        message:
          type: string
          description: The log line.
        timestamp:
          format: date-time
          type: string
          description: When the line was reported.
      required:
        - message
        - timestamp
    DeploymentSubStepDto:
      type: object
      properties:
        id:
          type: number
          description: The id of the sub-step.
        name:
          type: string
          description: The name of the sub-step.
        status:
          type: string
          description: The status of the sub-step.
          enum:
            - Pending
            - Running
            - Completed
            - Failed
        error:
          type: string
          description: The error, when the sub-step failed.
          nullable: true
        logs:
          description: The timestamped log lines reported for this sub-step.
          type: array
          items:
            $ref: "#/components/schemas/DeploymentStepLogDto"
        startedAt:
          format: date-time
          type: string
          description: When the sub-step was started.
          nullable: true
        completedAt:
          format: date-time
          type: string
          description: When the sub-step was completed or failed.
          nullable: true
      required:
        - id
        - name
        - status
        - error
        - logs
        - startedAt
        - completedAt
    DeploymentStepDto:
      type: object
      properties:
        resourceId:
          type: string
          description: The ID of the resource within the service definition.
        resourceName:
          type: string
          description: The display name of the resource.
        action:
          type: string
          description: Whether the resource is deployed or deleted in this step.
          enum:
            - Deploy
            - Delete
        status:
          type: string
          description: The status of the step.
          enum:
            - Pending
            - Running
            - Completed
            - Failed
        attempt:
          type: number
          description: The current attempt, greater than one when the step has been retried.
        maxAttempts:
          type: number
          description: The maximum number of attempts.
        error:
          type: string
          description: The error of the last failed attempt.
          nullable: true
        subSteps:
          description: The sub-steps as reported by the worker.
          type: array
          items:
            $ref: "#/components/schemas/DeploymentSubStepDto"
        logs:
          description: The timestamped log lines reported outside any sub-step.
          type: array
          items:
            $ref: "#/components/schemas/DeploymentStepLogDto"
        startedAt:
          format: date-time
          type: string
          description: When the step was started.
          nullable: true
        completedAt:
          format: date-time
          type: string
          description: When the step was completed or failed.
          nullable: true
      required:
        - resourceId
        - resourceName
        - action
        - status
        - attempt
        - maxAttempts
        - error
        - subSteps
        - logs
        - startedAt
        - completedAt
    DeploymentStepsDto:
      type: object
      properties:
        steps:
          description: The steps of the most recent deployment update.
          type: array
          items:
            $ref: "#/components/schemas/DeploymentStepDto"
      required:
        - steps
    ResourceInstanceLogDto:
      type: object
      properties:
        instanceId:
          type: string
          description: The identifier for instances or kubernetes deployments.
        messages:
          type: string
          description: The actual log message.
      required:
        - instanceId
        - messages
    ResourceLogsDto:
      type: object
      properties:
        resourceId:
          type: string
          description: The resource I.
        resourceType:
          type: string
          description: The type of the resource.
        resourceName:
          type: string
          description: The name of the resource.
        instances:
          description: The logs for the instances.
          type: array
          items:
            $ref: "#/components/schemas/ResourceInstanceLogDto"
      required:
        - resourceId
        - resourceType
        - resourceName
        - instances
    DeploymentLogsDto:
      type: object
      properties:
        resources:
          description: The resource.
          type: array
          items:
            $ref: "#/components/schemas/ResourceLogsDto"
      required:
        - resources
    DeploymentCheckSummaryDto:
      type: object
      properties:
        date:
          type: string
          description: The date for which the summary has been create.
        totalFailures:
          type: number
          description: The total number of failures on this dat.
        totalSuccesses:
          type: number
          description: The total number of successes on this dat.
      required:
        - date
        - totalFailures
        - totalSuccesses
    DeploymentCheckSummariesDto:
      type: object
      properties:
        checks:
          description: The summary per dat.
          type: array
          items:
            $ref: "#/components/schemas/DeploymentCheckSummaryDto"
      required:
        - checks
    DeploymentUsageSummaryDto:
      type: object
      properties:
        date:
          type: string
          description: The date for which the summary has been create.
        totalCores:
          type: number
          description: The total cores at the specified dat.
        totalMemoryGB:
          type: number
          description: The total memory at the specified date (in GB.
        totalVolumeGB:
          type: number
          description: The total volume at the specified date (in GB.
        totalStorageGB:
          type: number
          description: The total storage at the specified date (in GB.
      required:
        - date
        - totalCores
        - totalMemoryGB
        - totalVolumeGB
        - totalStorageGB
    DeploymentUsageSummariesDto:
      type: object
      properties:
        summaries:
          description: The usage summary per dat.
          type: array
          items:
            $ref: "#/components/schemas/DeploymentUsageSummaryDto"
      required:
        - summaries
...
    CreateDeploymentDto:
      type: object
      properties:
        name:
          type: string
          description: The optional name to describe the deployment.
          nullable: true
          maxLength: 100
        serviceId:
          type: number
          description: The ID of the service.
        parameters:
          type: object
          description: The parameters.
          additionalProperties: true
        confirmUrl:
          type: string
          description: The URL to call after the deployment has been created.
          nullable: true
        cancelUrl:
          type: string
          description: The URL to call after the deployment has been cancelled.
          nullable: true
      required:
        - name
        - serviceId
        - parameters
        - confirmUrl
        - cancelUrl
    DeploymentCreatedDto:
      type: object
      properties:
        deployment:
          description: The created deployment.
          nullable: true
          allOf:
            - $ref: "#/components/schemas/DeploymentDto"
        redirectUrl:
          type: string
          description: The redirect URL if the deployment cannot be completed automatically.
          nullable: true
      required:
        - deployment
        - redirectUrl
...